Add `package-lock.json` and so remove it from `.gitignore`
Package-lock to ensure a consistent install and compatible dependencies. Its purpose is to avoid situation where installing modules from the same package.json results in two different installs.
This change is not a breaking change due that older versions of NPM (< 5.x.x) will not use it.
IMO, this change is essential for going forward and make `adapt_authoring` code base and dependencies up-to-date in the future.

Related #1704
Related #1873