Add new config `dbConnectionUri`
It will allow developer to specific himself the whole "connection string". Indeed, there are some case that are not covered by the config, simple example: `ssl=true` that could be append to the query of the string.
I don't think adding a new config key for each case is a good idea.

A new dependency has been introduced `mongodb-uri`. Indeed, in case where `connect(db)` with a db, we need to change the database name in the connection string.